+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This page reports the main releases only and the main changes therein.
 - *align subsample* option added to persistent settings
 - double-click on processing colored box display opens processed frame
 - warning message appears when trying to load > 20 image frames
-- brush luminosity correction with slider in the main window
+- brush luminosity correction with slider in the main window and < and > shortcuts
 
 ### Fixed
 - fixed tooltip counter in the processing colored box display